add Open Camera Reel analytics
