+++ Explorer/Assets/DCL/PerformanceAndDiagnostics/Analytics/AnalyticsEvents.cs
@@ -58,13 +58,12 @@
             public const string TAKE_PHOTO = "take_photo";
 
             public const string CAMERA_REEL_OPEN = "open_camera_reel";
+
             public const string OPEN_PHOTO = "open_photo";
             public const string SHARE_PHOTO = "share_photo";
             public const string DOWNLOAD_PHOTO = "download_photo";
             public const string DELETE_PHOTO = "delete_photo";
-
             public const string PHOTO_TO_MARKETPLACE = "photo_to_marketplace";
-
             public const string PHOTO_JUMP_TO = "photo_jump_to";
         }
