+++ Explorer/Assets/DCL/PerformanceAndDiagnostics/Analytics/DCL.Analytics.asmdef
@@ -29,7 +29,8 @@
         "GUID:8baf705856414dad9a73b3f382f1bc8b",
         "GUID:254c118155004a75a699a9410d44311f",
         "GUID:ca926b7b13255b84199c21c5e2d9dfa4",
-        "GUID:ecd9230b77a12534cbb829d1b5fcd665"
+        "GUID:ecd9230b77a12534cbb829d1b5fcd665",
+        "GUID:f82471a6db95848d88faad2ceb31c9c9"
     ],
     "includePlatforms": [],
     "excludePlatforms": [],
